Restructures the listeners removal test scenario
