+++ test/msw-api/setup-worker/stop/removes-all-listeners.mocks.ts
@@ -1,11 +1,12 @@
 import { setupWorker, rest } from 'msw'
 
-const createWorker = () =>
-  setupWorker(
+const createWorker = () => {
+  return setupWorker(
     rest.get('/user', (req, res, ctx) => {
       return res(ctx.status(200))
     }),
   )
+}
 
 // @ts-ignore
 window.__MSW_CREATE_WORKER__ = createWorker
